Move NUnitConfiguration to Helpers namespace
