+++ src/TestEngine/nunit.engine.core/CoreEngine.cs
@@ -1,4 +1,4 @@
-﻿// ***********************************************************************
+// ***********************************************************************
 // Copyright (c) 2011 Charlie Poole, Rob Prouse
 //
 // Permission is hereby granted, free of charge, to any person obtaining
@@ -26,6 +26,7 @@ using System.Diagnostics;
 using System.IO;
 using System.Reflection;
 using System.Runtime.InteropServices;
+using NUnit.Engine.Helpers;
 using NUnit.Engine.Internal;
 using NUnit.Engine.Services;
